+++ spec/arachni/rpc/em/client_spec.rb
@@ -11,8 +11,8 @@ describe Arachni::RPC::EM::Client do
     it 'retains stability and consistency under heavy load' do
         client = start_client( rpc_opts )
 
-        n    = 100_000
-        cnt  = 0
+        n   = 100_000
+        cnt = 0
 
         mismatches = []
 
@@ -32,8 +32,89 @@ describe Arachni::RPC::EM::Client do
 
     describe '#initialize' do
         it 'should be able to properly assign class options (including :role)' do
-            opts = rpc_opts.merge( :role => :client )
+            opts = rpc_opts.merge( role: :client )
             start_client( opts ).opts.should == opts
+        end
+
+        context 'when passed no connection information' do
+            it 'raises ArgumentError' do
+                begin
+                    described_class.new({})
+                rescue => e
+                    e.should be_kind_of ArgumentError
+                end
+            end
+        end
+
+        describe 'option' do
+            describe :socket do
+                it 'connects to it' do
+                    client = start_client( rpc_opts_with_socket )
+                    client.call( 'test.foo', 1 ).should == 1
+                end
+
+                it 'retains stability and consistency under heavy load' do
+                    client = start_client( rpc_opts_with_socket )
+
+                    n    = 100_000
+                    cnt  = 0
+
+                    mismatches = []
+
+                    n.times do |i|
+                        arg = 'a' * i
+                        client.call( 'test.foo', arg ) do |res|
+                            cnt += 1
+                            mismatches << [i, arg, res] if arg != res
+                            ::EM.stop if cnt == n || mismatches.any?
+                        end
+                    end
+
+                    Arachni::RPC::EM.block
+                    cnt.should > 0
+                    mismatches.should be_empty
+                end
+
+                context 'when passed an invalid socket path' do
+                    it 'raises ArgumentError' do
+                        begin
+                            described_class.new( socket: 'blah' )
+                        rescue => e
+                            e.should be_kind_of ArgumentError
+                        end
+                    end
+                end
+            end
+        end
+
+        context 'when passed a host but not a port' do
+            it 'raises ArgumentError' do
+                begin
+                    described_class.new( host: 'test' )
+                rescue => e
+                    e.should be_kind_of ArgumentError
+                end
+            end
+        end
+
+        context 'when passed a port but not a host' do
+            it 'raises ArgumentError' do
+                begin
+                    described_class.new( port: 9999 )
+                rescue => e
+                    e.should be_kind_of ArgumentError
+                end
+            end
+        end
+
+        context 'when passed an invalid port' do
+            it 'raises ArgumentError' do
+                begin
+                    described_class.new( host: 'tt', port: 'blah' )
+                rescue => e
+                    e.should be_kind_of ArgumentError
+                end
+            end
         end
     end